requirements.txt
numpy==1.20.2
streamlit==0.80.0
tensorflow==2.3.0
pandas==1.2.4
Pillow==8.2.0
